+++ scripts/terminalserver/terminalserver.jl
@@ -2,12 +2,14 @@
 let
     args = [popfirst!(Base.ARGS) for _ in 1:8]
     conn_pipename, debug_pipename, telemetry_pipename = args[1:3]
+    has_revise = true
 
     include("load_vscodeserver.jl")
 
     # load Revise ?
     if "USE_REVISE=true" in args
         try
+            # Backward compatiblity for older julia versions
             @static if VERSION ≥ v"1.5"
                 using Revise
             else
@@ -15,6 +17,7 @@ let
                 Revise.async_steal_repl_backend()
             end
         catch err
+            has_revise = false
         end
     end
 
@@ -28,4 +31,7 @@ let
     end
 
     VSCodeServer.serve(conn_pipename, debug_pipename; is_dev="DEBUG_MODE=true" in args, error_handler = (err, bt) -> VSCodeServer.global_err_handler(err, bt, telemetry_pipename, "REPL"))
+    if !has_revise
+        VSCodeServer.JSONRPC.send_notification(VSCodeServer.conn_endpoint[], "norevise", has_revise)
+    end
 end